fix: make `Provider` and `Signer` interfaces, update `Wallet` to be non-abstract, and remove `LocalWallet`
