+++ src/index.js
@@ -3,7 +3,6 @@ import Logger from "js-logger";
 export * from "./exports.js";
 
 export { default as LocalProvider } from "./LocalProvider.js";
-export { default as LocalWallet } from "./LocalWallet.js";
 export { default as Client } from "./client/NodeClient.js";
 
 if (
